Learning JWT

+++ Mern Projects/goalsetter-app/backend/routes/goalRoutes.js
@@ -1,13 +1,15 @@
 import express from "express";
 import Goal from "../model/goalModel.js";
+import { protect } from "../middleware/authMiddleware.js";
+import User from "../model/userModel.js";
 
 const router = express.Router();
 
-router.get("/", async (req, res) => {
+router.get("/", protect, async (req, res) => {
   const goals = await Goal.find();
   res.json(goals);
 });
-router.post("/", async (req, res) => {
+router.post("/", protect, async (req, res) => {
   if (!req.body.text) {
     res.status(400);
     throw new Error("Please add a text field.");
@@ -15,16 +17,27 @@ router.post("/", async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.json(goal);
 });
-router.put("/:id", async (req, res) => {
+router.put("/:id", protect, async (req, res) => {
   const goal = await Goal.findById(req.params.id);
 
   if (!goal) {
     res.status(400);
     throw new error("Goal not found");
+  }
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -33,14 +46,23 @@ router.put("/:id", async (req, res) => {
 
   res.json(updatedGoal);
 });
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", protect, async (req, res) => {
   const goal = await Goal.findById(req.params.id);
 
   if (!goal) {
     res.status(400);
     throw new error("Goal not found");
   }
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
 
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
   await Goal.findByIdAndDelete(req.params.id);
 
   res.json({ message: `deleted Goal ${req.params.id}` });
